+++ CHANGELOG.md
@@ -10,9 +10,9 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ### Added
 
+- Added optional input functions for Gene Annotation
 - Added optional input functions for RiPPs.
-- Added functions to RiPP class: cyclic, subclass, peptidases, precursor peptides
-- Added support for SMILES, compound synonym names, compound evidence, biological activities, biological targets
+- Added optional input functions for compounds.
 
 ### Changed